requirements_bot.txt
aiogram==3.10.0
python-dotenv==1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-dateutil>=2.8.2
